requirements.txt
django==1.7.3 
django-mptt==0.6.1
